feat: 로또 티켓 기능 추가
* 로또 티켓 생성 기능을 로또 번호 풀에서 로또 티켓으로 이동
* 로또 번호 풀 삭제
