+++ README.md
@@ -26,27 +26,23 @@
                 * [x] 지불 (로또 가격) -> 로또 개수
             * [x] 로또 단말기
                 * [x] 로또 생성 (로또 개수) -> 로또 티켓 목록
-                * [x] 로또 번호 풀
-                    * [x] 생성 (로또 개수) -> 로또 티켓
-                    * [x] 로또 티켓
-                        * [x] 로또
-                            * [x] 로또 번호는 6개
-                            * [x] 로또 번호 목록
-                                * [x] 로또 번호는 오름차순 정렬
-                                * [x] 로또 번호
-    * [ ] System 2. 사용자에게 **로또 게임**으로 로또 개수와 로또 번호 목록을 보여준다.
-        * [x] 로또 게임
-            * [x] 로또 개수
-            * [x] 로또 번호 목록
+                * [x] 로또 티켓
+                    * [x] 로또 번호 목록
+                        * [x] 로또 번호 오름차순 정렬
+                        * [x] 로또 번호 6개
+                        * [x] 로또 번호 랜덤 생성
+                        * [x] 로또 번호
+                            * [x] 숫자
+                            * [x] 1 ~ 45 사이
+    * [ ] System 2. 사용자에게 **로또 게임**으로 로또 개수와 로또 번호 목록을 보여준다. * [x] 로또 게임 * [x] 로또 개수 * [x] 로또 번호 목록
 * [ ] Step 2. 사용자에게 지난 주 **당첨 번호** 입력을 요청한다.
     * [ ] User 1. **당첨 번호**를 입력한다.
         * [ ] 로또 사업자
             * [ ] 당첨 번호
-                * [ ] 로또 번호 목록
-                    * [ ] 로또 번호는 중복 불가
-                    * [ ] 로또 번호
-                        * [ ] 숫자
-                        * [ ] 1 ~ 45 사이
+                * [x] 로또 티켓
+                    * [x] 로또 번호는 중복 불가
+                    * [x] 로또 번호
+                        * [x] 캐싱
     * [ ] System 1. **로또 사업자**에게 **로또 목록**의 **로또 당첨 결과** *확인*을 요청한다.
         * [ ] 로또 사업자
             * [ ] 로또 등수
